Looks and stability
With this version, user could now use it with microscope. The only missing functionality is detecting when the maximum volume has been reached and the pump need to start withdrawing instead of infusing.

User now inputs volume instead of rate (rate is calculated behind the scenes and sent to the pump).
User can now pause the observer without error. This was fixed by a using globals, might switch to something else in the future.
The GUI has been rinsed of earlier unusable or unnecessary things.